added putptr

+++ ft_print_helpers.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_print_helpers.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bea_s <bea_s@student.42.fr>                +#+  +:+       +#+        */
+/*   By: bsurilla <bsurilla@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/08 00:41:19 by bsurilla          #+#    #+#             */
-/*   Updated: 2026/06/08 00:45:13 by bea_s            ###   ########.fr       */
+/*   Updated: 2026/06/08 20:40:01 by bsurilla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,9 +46,9 @@ int	ft_putnbr(int n)
 		i++;
 	}
 	if (nb >= 10)
-	{	
+	{
 		i = i + ft_putnbr(nb / 10);
-	}	 
+	}
 	digit = nb % 10 + '0';
 	write(1, &digit, 1);
 	i++;
@@ -62,30 +62,30 @@ int	ft_putnbr_u(unsigned int n)
 
 	i = 0;
 	if (n >= 10)
-	{	
+	{
 		i = i + ft_putnbr_u(n / 10);
-	}	 
+	}
 	digit = n % 10 + '0';
 	write(1, &digit, 1);
 	i++;
 	return (i);
 }
 
-int	ft_putnbr_hex(unsigned int n, char *hex)
+int	ft_putnbr_hex(unsigned long n, char *hex)
 {
 	char		digit;
 	int			i;
 
 	i = 0;
 	if (n >= 16)
-	{	
+	{
 		i = i + ft_putnbr_hex(n / 16, hex);
-	}	 
+	}
 	digit = hex[n % 16];
 	write(1, &digit, 1);
 	i++;
 	return (i);
-} 
+}
 /* 
  #include <stdio.h>
  #include <limits.h>
@@ -104,4 +104,4 @@ int	ft_putnbr_hex(unsigned int n, char *hex)
 	printf(" is your number!\n it has %d digits.\n", ft_putnbr(n));	
 	printf(" is your number!\n it has %d digits.\n", ft_putnbr_u(u));
 	printf(" is your number!\n it has %d digits.\n", ft_putnbr_hex(u,hexbase));	
-}  */
+} */
